Move `Additive` to its own module
